discounts added

+++ src/main/java/com/booleanuk/core/InventoryItem.java
@@ -2,11 +2,11 @@ package com.booleanuk.core;
 
 public class InventoryItem {
     public String sku;
-    public Double price;
+    public int price;
     public String name;
     public String variant;
 
-    public InventoryItem(String sku,double price,String name, String variant) {
+    public InventoryItem(String sku,int price,String name, String variant) {
         this.sku = sku;
         this.price = price;
         this.name = name;
